+++ src/main/java/shadows/apotheosis/ench/library/EnchLibraryScreen.java
@@ -4,21 +4,27 @@ import java.text.DecimalFormat;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Locale;
 
 import javax.annotation.Nullable;
 
+import org.lwjgl.glfw.GLFW;
+
+import com.google.common.base.Strings;
+import com.mojang.blaze3d.platform.InputConstants;
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
 
 import it.unimi.dsi.fastutil.objects.Object2IntMap.Entry;
 import net.minecraft.ChatFormatting;
+import net.minecraft.client.gui.components.EditBox;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.client.renderer.GameRenderer;
 import net.minecraft.client.resources.language.I18n;
 import net.minecraft.client.resources.sounds.SimpleSoundInstance;
 import net.minecraft.network.chat.Component;
-import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.network.chat.FormattedText;
 import net.minecraft.network.chat.Style;
 import net.minecraft.network.chat.TextColor;
 import net.minecraft.network.chat.TextComponent;
@@ -44,6 +50,7 @@ public class EnchLibraryScreen extends AbstractContainerScreen<EnchLibraryContai
 	protected int startIndex;
 
 	protected List<LibrarySlot> data = new ArrayList<>();
+	protected EditBox filter = null;
 
 	public EnchLibraryScreen(EnchLibraryContainer container, Inventory inv, Component title) {
 		super(container, inv, title);
@@ -54,6 +61,26 @@ public class EnchLibraryScreen extends AbstractContainerScreen<EnchLibraryContai
 		this.inventoryLabelY = 149;
 		this.containerChanged();
 		container.setNotifier(this::containerChanged);
+	}
+
+	@Override
+	protected void init() {
+		super.init();
+		this.filter = this.addRenderableWidget(new EditBox(font, this.getGuiLeft() + 91, this.getGuiTop() + 20 + font.lineHeight + 2, 78, font.lineHeight + 4, filter, new TextComponent("")));
+		this.filter.setResponder(t -> this.containerChanged());
+	}
+
+	@Override
+	public boolean keyPressed(int pKeyCode, int pScanCode, int pModifiers) {
+		InputConstants.Key mouseKey = InputConstants.getKey(pKeyCode, pScanCode);
+		if (pKeyCode == GLFW.GLFW_KEY_ESCAPE && this.getFocused() == this.filter) {
+			this.setFocused(null);
+			this.filter.setFocus(false);
+			return true;
+		} else if (this.minecraft.options.keyInventory.isActiveAndMatches(mouseKey) && this.getFocused() == this.filter) {
+			return true;
+		}
+		return super.keyPressed(pKeyCode, pScanCode, pModifiers);
 	}
 
 	@Override
@@ -68,8 +95,14 @@ public class EnchLibraryScreen extends AbstractContainerScreen<EnchLibraryContai
 		super.renderTooltip(stack, mouseX, mouseY);
 		LibrarySlot libSlot = this.getHoveredSlot(mouseX, mouseY);
 		if (libSlot != null) {
-			List<MutableComponent> list = new ArrayList<>();
+			List<FormattedText> list = new ArrayList<>();
 			list.add(new TranslatableComponent(libSlot.ench.getDescriptionId()).setStyle(Style.EMPTY.withColor(TextColor.fromRgb(0xFFFF80)).setUnderlined(true)));
+			if (I18n.exists(libSlot.ench.getDescriptionId() + ".desc")) {
+				Component txt = new TranslatableComponent(libSlot.ench.getDescriptionId() + ".desc").setStyle(Style.EMPTY.withColor(ChatFormatting.GRAY).withItalic(true));
+				list.addAll(this.font.getSplitter().splitLines(txt, this.getGuiLeft() - 16, txt.getStyle()));
+				list.add(new TextComponent(""));
+			}
+
 			list.add(new TranslatableComponent("tooltip.enchlib.max_lvl", new TranslatableComponent("enchantment.level." + libSlot.maxLvl)).withStyle(ChatFormatting.GRAY));
 			list.add(new TranslatableComponent("tooltip.enchlib.points", format(libSlot.points), format(this.menu.getPointCap())).withStyle(ChatFormatting.GRAY));
 			list.add(new TextComponent(""));
@@ -84,7 +117,7 @@ public class EnchLibraryScreen extends AbstractContainerScreen<EnchLibraryContai
 				list.add(new TranslatableComponent("tooltip.enchlib.extracting", new TranslatableComponent("enchantment.level." + targetLevel)).withStyle(ChatFormatting.BLUE));
 				list.add(new TranslatableComponent("tooltip.enchlib.cost", cost).withStyle(cost > libSlot.points ? ChatFormatting.RED : ChatFormatting.GOLD));
 			}
-			this.renderComponentTooltip(stack, list, this.getGuiLeft() - 16 - this.font.width(list.get(2)), mouseY, this.font);
+			this.renderComponentTooltip(stack, list, this.getGuiLeft() - 16 - list.stream().map(this.font::width).max(Integer::compare).get(), mouseY, this.font);
 		}
 	}
 
@@ -104,6 +137,9 @@ public class EnchLibraryScreen extends AbstractContainerScreen<EnchLibraryContai
 			this.renderEntry(stack, this.data.get(idx), this.leftPos + 8, this.topPos + 14 + 19 * (idx - this.startIndex), mouseX, mouseY);
 			idx++;
 		}
+
+		this.font.draw(stack, new TranslatableComponent("tooltip.enchlib.nfilt"), this.getGuiLeft() + 91, this.getGuiTop() + 20, 4210752);
+		this.font.draw(stack, new TranslatableComponent("tooltip.enchlib.ifilt"), this.getGuiLeft() + 91, this.getGuiTop() + 50, 4210752);
 	}
 
 	private void renderEntry(PoseStack stack, LibrarySlot data, int x, int y, int mouseX, int mouseY) {
@@ -183,7 +219,8 @@ public class EnchLibraryScreen extends AbstractContainerScreen<EnchLibraryContai
 
 	private void containerChanged() {
 		this.data.clear();
-		for (Entry<Enchantment> e : this.menu.getPointsForDisplay()) {
+		List<Entry<Enchantment>> entries = filter(this.menu.getPointsForDisplay());
+		for (Entry<Enchantment> e : entries) {
 			this.data.add(new LibrarySlot(e.getKey(), e.getIntValue(), this.menu.getMaxLevel(e.getKey())));
 		}
 
@@ -192,6 +229,21 @@ public class EnchLibraryScreen extends AbstractContainerScreen<EnchLibraryContai
 			this.startIndex = 0;
 		}
 		Collections.sort(this.data, (a, b) -> I18n.get(a.ench.getDescriptionId()).compareTo(I18n.get(b.ench.getDescriptionId())));
+	}
+
+	private List<Entry<Enchantment>> filter(List<Entry<Enchantment>> list) {
+		return list.stream().filter(this::isAllowedByItem).filter(this::isAllowedBySearch).toList();
+	}
+
+	private boolean isAllowedByItem(Entry<Enchantment> e) {
+		ItemStack stack = this.menu.ioInv.getItem(2);
+		return stack.isEmpty() || stack.canApplyAtEnchantingTable(e.getKey());
+	}
+
+	private boolean isAllowedBySearch(Entry<Enchantment> e) {
+		String name = I18n.get(e.getKey().getDescriptionId()).toLowerCase(Locale.ROOT);
+		String search = this.filter == null ? "" : this.filter.getValue().trim().toLowerCase(Locale.ROOT);
+		return Strings.isNullOrEmpty(search) || ChatFormatting.stripFormatting(name).contains(search);
 	}
 
 	@Nullable